requirements.txt
cupy-cuda11x
scikit-learn
numpy
